+++ R/ari_stitch.R
@@ -18,22 +18,43 @@
 #' @param images A vector of paths to images.
 #' @param audio A list of \code{Wave}s from tuneR.
 #' @param output A path to the video file which will be created.
+#' @param verbose print diagnostic messages.  If > 1, then more are printed
+#' @param cleanup If \code{TRUE}, interim files are deleted
+#' @param ffmpeg_opts additional options to send to \code{ffmpeg}.
+#' This is an advanced option, use at your own risk
+#' @param divisible_height Make height divisible by 2, which may 
+#' be required if getting "height not divisible by 2" error.
+#' @param audio_codec The audio encoder for the splicing.  If this
+#' fails, try \code{copy}.
+#' @param video_codec The video encoder for the splicing.  If this
+#' fails, see \code{ffmpeg -codecs}
+#' @param audio_bitrate Bit rate for audio. Passed to \code{-b:a}.
+#' @param video_bitrate Bit rate for video. Passed to \code{-b:v}.
+#' 
+#' @return A logical value, with the attribute \code{outfile} for the
+#' output file.
+
 #' @importFrom purrr reduce discard
 #' @importFrom tuneR bind writeWave
 #' @export
 #' @examples 
-#' \dontrun{
-#' 
-#' library(tuneR)
-#' library(purrr)
-#' 
-#' slides <- c("intro.jpeg", "equations.jpeg", "questions.jpeg")
-#' sound <- map(c("rec1.wav", "rec2.wav", "rec3.wav"), readWave)
-#' 
-#' ari_stitch(slides, sound)
-#' 
+#' if (ffmpeg_version_sufficient()) {
+#' result = ari_stitch(
+#' ari_example(c("mab1.png", "mab2.png")),
+#' list(tuneR::noise(), tuneR::noise()))
 #' }
-ari_stitch <- function(images, audio, output = "output.mp4"){
+ari_stitch <- function(
+  images, audio, 
+  output = tempfile(fileext = ".mp4"),
+  verbose = FALSE,
+  cleanup = TRUE,
+  ffmpeg_opts = "",
+  divisible_height = FALSE,
+  audio_codec = get_audio_codec(),
+  video_codec = get_video_codec(),
+  audio_bitrate = "192k",
+  video_bitrate = NULL
+){
   stopifnot(length(images) > 0)
   images <- normalizePath(images)
   output_dir <- normalizePath(dirname(output))
@@ -43,30 +64,101 @@ ari_stitch <- function(images, audio, output = "output.mp4"){
     all(file.exists(images)),
     dir.exists(output_dir)
   )
+  if (is.character(audio)) {
+    audio = lapply(audio, tuneR::readMP3)
+    audio = lapply(audio, function(wav) {
+      ideal_duration <- ceiling(length(wav@left) / wav@samp.rate)
+      left = rep(0, 
+                 wav@samp.rate * ideal_duration - length(wav@left))
+      right = numeric(0)
+      if (wav@stereo) {
+        right = left
+      }
+      end_wav = tuneR::Wave(
+        left = left,
+        right = right,
+        bit = wav@bit, samp.rate = wav@samp.rate)         
+      wav <- bind(wav, end_wav)
+      wav      
+    })
+  }
+  # Make a hard path
+  output = file.path(output_dir, basename(output))
   
-  wav <- reduce(audio, bind)
+  if (verbose > 0) {
+    message("Writing out Wav for audio")
+  }
+  wav <- purrr::reduce(audio, bind)
   wav_path <- file.path(output_dir, paste0("ari_audio_", grs(), ".wav"))
   writeWave(wav, filename = wav_path)
-  on.exit(unlink(wav_path, force = TRUE), add = TRUE)
-  
-  input_txt_path <- file.path(output_dir, paste0("ari_input_", grs(), ".txt"))
-  for(i in 1:length(images)){
-    cat(paste0("file ", "'", images[i], "'", "\n"), file = input_txt_path, append = TRUE)
-    cat(paste0("duration ", duration(audio[[i]]), "\n"), file = input_txt_path, append = TRUE)
-  }
-  cat(paste0("file ", "'", images[i], "'", "\n"), file = input_txt_path, append = TRUE)
-  
-  ffmpeg <- discard(c(Sys.getenv("ffmpeg"), Sys.which("ffmpeg")), ~ nchar(.x) == 0)[1]
-  
-  if(is.na(ffmpeg)){
-    stop("Could not find ffmpeg. See the documentation for ari_stitch() for more details.")
+  if (cleanup) {
+    on.exit(unlink(wav_path, force = TRUE), add = TRUE)
   }
   
-  command <- paste(ffmpeg, "-y -f concat -safe 0 -i", input_txt_path, "-i", 
-                   wav_path, "-c:v libx264 -c:a aac -b:a 192k -shortest -vsync vfr -pix_fmt yuv420p",
-                   output)
-  system(command)
+  input_txt_path <-  paste0("ari_input_", 
+                                     grs(), 
+                                     ".txt")
+  ## on windows ffmpeg cancats names adding the working directory, so if
+  ## complete url is provided it adds it twice.
+  # if (.Platform$OS.type == "windows") {
+  #   images <- basename(images)   
+  # }
+  for (i in seq_along(images)) {
+    cat(paste0("file ", "'", images[i], "'", "\n"), 
+        file = input_txt_path, append = TRUE)
+    cat(paste0("duration ", duration(audio[[i]]), "\n"), 
+        file = input_txt_path, append = TRUE)
+  }
+  cat(paste0("file ", "'", images[i], "'", "\n"), 
+      file = input_txt_path, append = TRUE)
+  # needed for users as per 
+  # https://superuser.com/questions/718027/
+  # ffmpeg-concat-doesnt-work-with-absolute-path
+  # input_txt_path = normalizePath(input_txt_path, winslash = "\\")
   
-  on.exit(unlink(input_txt_path, force = TRUE), add = TRUE)
-  invisible(file.exists(output) && file.size(output) > 0)
+  ffmpeg = ffmpeg_exec(quote = TRUE)
+  
+  if (divisible_height) {
+    ffmpeg_opts = c(ffmpeg_opts, 
+                    '-vf "scale=trunc(iw/2)*2:trunc(ih/2)*2"')
+  }
+  
+  ffmpeg_opts = paste(ffmpeg_opts, collapse = " ")
+  # shQuote should seankross/ari#5
+  command <- paste(
+    ffmpeg, "-y -f concat -safe 0 -i", shQuote(input_txt_path), 
+    "-i", shQuote(wav_path), 
+    ifelse(!is.null(video_codec), paste("-c:v", video_codec),
+           ""),
+    ifelse(!is.null(audio_codec), paste("-c:a", audio_codec),
+           ""),    
+    ifelse(!is.null(audio_bitrate), paste("-b:a", audio_bitrate),
+           ""), 
+    ifelse(!is.null(video_bitrate), paste("-b:v", video_bitrate),
+           ""), 
+    " -shortest -vsync vfr -pix_fmt yuv420p",
+    ffmpeg_opts,
+    shQuote(output))
+  if (verbose > 0) {
+    message(command)
+  }
+  if (verbose > 1) {
+    message("Input text path is:")
+    cat(readLines(input_txt_path), sep = "\n")
+  }
+  res = system(command)
+  if (res != 0) {
+    warning("Result was non-zero for ffmpeg")
+  }
+  
+  if (cleanup) {
+    on.exit(unlink(input_txt_path, force = TRUE), add = TRUE)
+  }
+  res = file.exists(output) && file.size(output) > 0
+  if (!cleanup) {
+    attr(res, "txt_path") = input_txt_path
+    attr(res, "wav_path") = wav_path
+  }
+  attr(res, "outfile") = output
+  invisible(res)
 }