+++ tests/testthat/test_ari_narrate.R
@@ -1,21 +1,48 @@
 context("Test ari_narrate()")
 
+
+if (ffmpeg_version_sufficient()) {
+  res = ffmpeg_audio_codecs()
+  if (is.null(res)) {
+    fdk_enabled = FALSE
+  } else {
+    fdk_enabled = grepl("fdk", res[ res$codec == "aac", "codec_name"])
+  }  
+} else {
+  fdk_enabled = FALSE
+}
+if (fdk_enabled) {
+  audio_codec = "libfdk_aac"
+} else {
+  audio_codec = "ac3"
+}
+
 skip_narrate <- function(){
-  if(Sys.getenv("SKIP_NARRATE") != ""){
+  if (Sys.getenv("SKIP_NARRATE") != "") {
     skip("skipping ari_narrate()")
   }
 }
 
 video <- file.path(tempdir(), "output.mp4")
 #video <- file.path(getwd(), "output.mp4")
+if (nzchar(Sys.getenv("AWS_ACCESS_KEY_ID"))) {
+  run_voice = aws.polly::list_voices()$Id[1]
+} else {
+  run_voice = "Joanna"
+}
 
 test_that("Ari can make a video from local HTML slides.", {
   skip_on_cran()
   skip_narrate()
-  
+  skip_amazon_not_authorized()
+
   ari_narrate(system.file("test", "ari_intro_script.md", package = "ari"),
               system.file("test", "ari_intro.html", package = "ari"),
-              video, voice = list_voices()$Id[1], capture_method = "iterative")
+              video, voice = run_voice,
+              capture_method = "iterative",
+              service = "amazon",
+              verbose = TRUE,
+              audio_codec = audio_codec)
   expect_true(file.size(video) > 50000)
 })
 
@@ -24,22 +51,29 @@ unlink(video, force = TRUE)
 test_that("Ari can make a video from HTML slides on the web.", {
   skip_on_cran()
   skip_narrate()
-  
+  skip_amazon_not_authorized()
+
   ari_narrate(system.file("test", "ari_intro_script.md", package = "ari"),
-              "http://seankross.com/ari/inst/test/ari_intro.html",
-              video, voice = list_voices()$Id[1], capture_method = "iterative")
+              "https://seankross.com/ari/inst/test/ari_intro.html",
+              video, voice = run_voice, 
+              capture_method = "iterative",
+              verbose = TRUE,
+              audio_codec = audio_codec)
   expect_true(file.size(video) > 50000)
 })
 
 unlink(video, force = TRUE)
 
-test_that("Ari use an Rmd file with HTML comments for a script.", {
+test_that("Ari can use an Rmd file with HTML comments for a script.", {
   skip_on_cran()
   skip_narrate()
-  
+  skip_amazon_not_authorized()
+
   ari_narrate(system.file("test", "ari_comments.Rmd", package = "ari"),
               system.file("test", "ari_intro.html", package = "ari"),
-              video, voice = list_voices()$Id[1], capture_method = "iterative")
+              video, voice = run_voice, capture_method = "iterative",
+              verbose = TRUE,
+              audio_codec = audio_codec)
   expect_true(file.size(video) > 50000)
 })